feat: Add URL flag to print JIRA ticket URL
This commit adds a new flag, `--url` or `-u`, to the root command. When this flag is provided, the command will print the JIRA ticket URL based on the configured JIRA URL and the provided ticket number. Additionally, if the `--copy` or `-c` flag is also provided, the URL will be copied to the clipboard. This feature enhances the functionality of the command by allowing users to quickly access the JIRA ticket associated with a specific task.